# Optimize defining methods on deeply included modules