--- a/vm.c
+++ b/vm.c
@@ -1557,6 +1557,7 @@ rb_iter_break_value(VALUE val)
 /* optimization: redefine management */
 
 static st_table *vm_opt_method_table = 0;
+static st_table *vm_opt_mid_table = 0;
 
 static int
 vm_redefinition_check_flag(VALUE klass)
@@ -1574,6 +1575,16 @@ vm_redefinition_check_flag(VALUE klass)
     if (klass == rb_cFalseClass) return FALSE_REDEFINED_OP_FLAG;
     if (klass == rb_cProc) return PROC_REDEFINED_OP_FLAG;
     return 0;
+}
+
+int
+rb_vm_check_optimizable_mid(VALUE mid)
+{
+    if (!vm_opt_mid_table) {
+      return FALSE;
+    }
+
+    return st_lookup(vm_opt_mid_table, mid, NULL);
 }
 
 static int
@@ -1630,6 +1641,7 @@ add_opt_method(VALUE klass, ID mid, VALUE bop)
 
     if (me && vm_redefinition_check_method_type(me->def)) {
 	st_insert(vm_opt_method_table, (st_data_t)me, (st_data_t)bop);
+	st_insert(vm_opt_mid_table, (st_data_t)mid, (st_data_t)Qtrue);
     }
     else {
 	rb_bug("undefined optimized method: %s", rb_id2name(mid));
@@ -1643,6 +1655,7 @@ vm_init_redefined_flag(void)
     VALUE bop;
 
     vm_opt_method_table = st_init_numtable();
+    vm_opt_mid_table = st_init_numtable();
 
 #define OP(mid_, bop_) (mid = id##mid_, bop = BOP_##bop_, ruby_vm_redefined_flag[bop] = 0)
 #define C(k) add_opt_method(rb_c##k, mid, bop)