--- a/vm_method.c
+++ b/vm_method.c
@@ -70,7 +70,7 @@ rb_class_clear_method_cache(VALUE klass, VALUE arg)
     mjit_remove_class_serial(RCLASS_SERIAL(klass));
     RCLASS_SERIAL(klass) = rb_next_class_serial();
 
-    if (RB_TYPE_P(klass, T_ICLASS)) {
+    if (BUILTIN_TYPE(klass) == T_ICLASS) {
 	struct rb_id_table *table = RCLASS_CALLABLE_M_TBL(klass);
 	if (table) {
 	    rb_id_table_clear(table);