@jhawthorn 2019-12-17
This also seems to (very slightly) improve performance of Ruby startup (even with nothing loaded).

We can see the methods in question from `perf record --freq=32000 ./ruby --disable-gems -e ''`

<img width="534" alt="Screen Shot 2019-12-17 at 09 10 45" src="https://user-images.githubusercontent.com/131752/71018442-583c4f80-20ad-11ea-8a6c-11bed3d2fa32.png">

`perf stat` shows a significant reduction in cycles/instructions (but a relatively small change in runtime)

**Before**

```
$ make -j16 ruby && perf stat -r 500 ./ruby --disable-gems -e ''
        BASERUBY = /home/jhawthorn/.rubies/ruby-2.6.3/bin/ruby --disable=gems
        CC = gcc
        LD = ld
        LDSHARED = gcc -shared
        CFLAGS = -O3 -ggdb3 -Wall -Wextra -Werror=deprecated-declarations -Werror=duplicated-cond -Werror=implicit-function-declaration -Werror=implicit-int -Werror=misleading-indentation -Werror=pointer-arith -Werror=write-strings -Wimplicit-fallthrough=0 -Wmissing-noreturn
-Wno-cast-function-type -Wno-constant-logical-operand -Wno-long-long -Wno-missing-field-initializers -Wno-overlength-strings -Wno-packed-bitfield-compat -Wno-parentheses-equality -Wno-self-assign -Wno-tautological-compare -Wno-unused-parameter -Wno-unused-value -Wsuggest-attr
ibute=format -Wsuggest-attribute=noreturn -Wunused-variable -std=gnu99
        XCFLAGS = -D_FORTIFY_SOURCE=2 -fstack-protector-strong -fno-strict-overflow -DRUBY_DEVEL=1 -fvisibility=hidden -fexcess-precision=standard -DRUBY_EXPORT -fPIE -DCANONICALIZATION_FOR_MATHN -I. -I.ext/include/x86_64-linux -I./include -I. -I./enc/unicode/12.1.0
        CPPFLAGS =
        DLDFLAGS = -Wl,--compress-debug-sections=zlib -fstack-protector-strong -pie
        SOLIBS = -lz -lpthread -lrt -lrt -lgmp -ldl -lcrypt -lm
        LANG = en_US.UTF-8
        LC_ALL =
        LC_CTYPE = en_US.UTF-8
        MFLAGS = -j16 --jobserver-auth=4,5
gcc (GCC) 9.2.0
Copyright (C) 2019 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

make: 'ruby' is up to date.

 Performance counter stats for './ruby --disable-gems -e ' (500 runs):

              7.24 msec task-clock:u              #    0.970 CPUs utilized            ( +-  0.59% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             1,589      page-faults:u             #    0.219 M/sec                    ( +-  0.01% )
         8,872,760      cycles:u                  #    1.226 GHz                      ( +-  1.10% )  (59.19%)
           247,261      stalled-cycles-frontend:u #    2.79% frontend cycles idle     ( +-  0.57% )  (66.48%)
         1,052,212      stalled-cycles-backend:u  #   11.86% backend cycles idle      ( +-  0.19% )  (95.11%)
        15,544,068      instructions:u            #    1.75  insn per cycle
                                                  #    0.07  stalled cycles per insn  ( +-  0.00% )
         3,417,975      branches:u                #  472.144 M/sec                    ( +-  0.01% )
            76,959      branch-misses:u           #    2.25% of all branches          ( +-  0.46% )  (79.22%)

         0.0074641 +- 0.0000437 seconds time elapsed  ( +-  0.59% )
```


**After**

```
$ make -j16 ruby && perf stat -r 500 ./ruby --disable-gems -e ''
        BASERUBY = /home/jhawthorn/.rubies/ruby-2.6.3/bin/ruby --disable=gems
        CC = gcc
        LD = ld
        LDSHARED = gcc -shared
        CFLAGS = -O3 -ggdb3 -Wall -Wextra -Werror=deprecated-declarations -Werror=duplicated-cond -Werror=implicit-function-declaration -Werror=implicit-int -Werror=misleading-indentation -Werror=pointer-arith -Werror=write-strings -Wimplicit-fallthrough=0 -Wmissing-noreturn
-Wno-cast-function-type -Wno-constant-logical-operand -Wno-long-long -Wno-missing-field-initializers -Wno-overlength-strings -Wno-packed-bitfield-compat -Wno-parentheses-equality -Wno-self-assign -Wno-tautological-compare -Wno-unused-parameter -Wno-unused-value -Wsuggest-attr
ibute=format -Wsuggest-attribute=noreturn -Wunused-variable -std=gnu99
        XCFLAGS = -D_FORTIFY_SOURCE=2 -fstack-protector-strong -fno-strict-overflow -DRUBY_DEVEL=1 -fvisibility=hidden -fexcess-precision=standard -DRUBY_EXPORT -fPIE -DCANONICALIZATION_FOR_MATHN -I. -I.ext/include/x86_64-linux -I./include -I. -I./enc/unicode/12.1.0
        CPPFLAGS =
        DLDFLAGS = -Wl,--compress-debug-sections=zlib -fstack-protector-strong -pie
        SOLIBS = -lz -lpthread -lrt -lrt -lgmp -ldl -lcrypt -lm
        LANG = en_US.UTF-8
        LC_ALL =
        LC_CTYPE = en_US.UTF-8
        MFLAGS = -j16 --jobserver-auth=4,5
gcc (GCC) 9.2.0
Copyright (C) 2019 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

make: 'ruby' is up to date.

 Performance counter stats for './ruby --disable-gems -e ' (500 runs):

              6.86 msec task-clock:u              #    0.968 CPUs utilized            ( +-  0.64% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             1,588      page-faults:u             #    0.232 M/sec                    ( +-  0.01% )
         7,877,563      cycles:u                  #    1.148 GHz                      ( +-  1.21% )  (57.39%)
           237,588      stalled-cycles-frontend:u #    3.02% frontend cycles idle     ( +-  0.61% )  (68.23%)
         1,016,882      stalled-cycles-backend:u  #   12.91% backend cycles idle      ( +-  0.15% )  (97.05%)
        14,052,135      instructions:u            #    1.78  insn per cycle
                                                  #    0.07  stalled cycles per insn  ( +-  0.01% )
         2,964,832      branches:u                #  432.121 M/sec                    ( +-  0.01% )
            75,965      branch-misses:u           #    2.56% of all branches          ( +-  0.52% )  (77.33%)

         0.0070858 +- 0.0000450 seconds time elapsed  ( +-  0.63% )
```